Accept floats in ReadingsSanitizer, LevelsBoundary
Return float in TimeOfFlightLevelStrategy

+++ com/components/LevelsBoundary.py
@@ -1,5 +1,5 @@
 class LevelsBoundary(object):
 
-    def __init__(self, full_level: int, empty_level: int):
+    def __init__(self, full_level: float, empty_level: float):
         self.full_level = full_level
         self.empty_level = empty_level
